RT 67207 - Stop doing tied on globs (sprout)
I was mistaken. It is possible to call a method on a typeglob, as in *foo->bar. So this better
patch makes tied handles work in 5.6 and skips the special handling in 5.8 and higher, as it is
unnecessary.

I added a local *@, because of the new eval{}, which rendered the existing undef
unnecessary.

I used defined tied , because the object implementing the tied could have boolean
overloading that returns false (paranoid, I know).