Avoid clobbering PL_na
We used to do SvPV(string, PL_na) in a few places.  This is not correct as it
sets the interpreter variable PL_na to the length of the string.  Conceptually,
PL_na is just a readability-improving alias for 0.  So if anyone else using the
same interpreter then does newSVpv(other_string, PL_na), an invalid read
occurs.  The test suite did not catch this because we use newSVpv(other_string,
0) everywhere.

Simply use SvPV_nolen(string) to fix this.